Move request references to separate file
